Downloads/Vehicle Repair System/vehicle_service/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers==3.14.0
requests==2.28.2